feat(sidebar): move account into project picker, landing logo in footer
- The account section (user row + tier-gated upgrade + log out) now lives at
  the BOTTOM of the project picker dropdown (new components/projects/account-menu.tsx),
  matching the reference layout. Removes the standalone footer SidebarUserMenu.
- Footer shows the landing Laminar wordmark (square icon when collapsed),
  replacing the old icons/* logo.
